+++ src/constants/countryList.ts
@@ -1,871 +1,1086 @@
-import { CountryListType } from "../types/list";
-
+export type CountryListType = {
+  code: string;
+  krNation: string;
+  enNation: string;
+};
 export const countryList: CountryListType[] = [
   {
+    code: "GH",
     krNation: "가나",
     enNation: "Ghana",
   },
   {
+    code: "GA",
     krNation: "가봉",
     enNation: "Gabon",
   },
   {
+    code: "GU",
     krNation: "가이아나",
     enNation: "Guyana",
   },
   {
+    code: "GA",
     krNation: "감비아",
     enNation: "Gambia",
   },
   {
+    code: "GU",
     krNation: "괌",
     enNation: "Guam",
   },
   {
+    code: "GU",
     krNation: "과들루프",
     enNation: "Guadeloupe",
   },
   {
+    code: "GU",
     krNation: "과테말라",
     enNation: "Guatemala",
   },
   {
+    code: "GR",
     krNation: "그레나다",
     enNation: "Grenada",
   },
   {
+    code: "GR",
     krNation: "그리스",
     enNation: "Greece",
   },
   {
+    code: "GR",
     krNation: "그린란드",
     enNation: "Greenland",
   },
   {
+    code: "GU",
     krNation: "기니",
     enNation: "Guinea",
   },
   {
+    code: "GU",
     krNation: "기니비사우",
     enNation: "Guinea-Bissau",
   },
   {
+    code: "NA",
     krNation: "나미비아",
     enNation: "Namibia",
   },
   {
+    code: "NA",
     krNation: "나우루",
     enNation: "Nauru",
   },
   {
+    code: "NI",
     krNation: "나이지리아",
     enNation: "Nigeria",
   },
   {
+    code: "SO",
     krNation: "남수단",
     enNation: "South Sudan",
   },
   {
+    code: "SO",
     krNation: "남아프리카 공화국",
     enNation: "South Africa",
   },
   {
+    code: "TH",
     krNation: "네덜란드",
     enNation: "the Netherlands",
   },
   {
+    code: "NE",
     krNation: "네팔",
     enNation: "Nepal",
   },
   {
+    code: "NO",
     krNation: "노르웨이",
     enNation: "Norway",
   },
   {
+    code: "NE",
     krNation: "누벨칼레도니",
     enNation: "New Caledonia",
   },
   {
+    code: "NE",
     krNation: "뉴질랜드",
     enNation: "New Zealand",
   },
   {
+    code: "NI",
     krNation: "니우에",
     enNation: "Niue",
   },
   {
+    code: "NI",
     krNation: "니제르",
     enNation: "Niger",
   },
   {
+    code: "NI",
     krNation: "니카라과",
     enNation: "Nicaragua",
   },
   {
+    code: "RE",
     krNation: "대한민국",
     enNation: "Republic of Korea",
   },
   {
+    code: "DE",
     krNation: "덴마크",
     enNation: "Denmark",
   },
   {
+    code: "DO",
     krNation: "도미니카 공화국",
     enNation: "Dominican R.",
   },
   {
+    code: "DO",
     krNation: "도미니카 연방",
     enNation: "Dominica",
   },
   {
+    code: "GE",
     krNation: "독일",
     enNation: "Germany",
   },
   {
+    code: "EA",
     krNation: "동티모르",
     enNation: "East Timor",
   },
   {
+    code: "LA",
     krNation: "라오스",
     enNation: "Laos",
   },
   {
+    code: "LI",
     krNation: "라이베리아",
     enNation: "Liberia",
   },
   {
+    code: "LA",
     krNation: "라트비아",
     enNation: "Latvia",
   },
   {
+    code: "RU",
     krNation: "러시아",
     enNation: "Russia",
   },
   {
+    code: "LE",
     krNation: "레바논",
     enNation: "Lebanon",
   },
   {
+    code: "LE",
     krNation: "레소토",
     enNation: "Lesotho",
   },
   {
+    code: "RO",
     krNation: "루마니아",
     enNation: "Romania",
   },
   {
+    code: "LU",
     krNation: "룩셈부르크",
     enNation: "Luxembourg",
   },
   {
+    code: "RW",
     krNation: "르완다",
     enNation: "Rwanda",
   },
   {
+    code: "LI",
     krNation: "리비아",
     enNation: "Libya",
   },
   {
+    code: "LI",
     krNation: "리투아니아",
     enNation: "Lithuania",
   },
   {
+    code: "LI",
     krNation: "리히텐슈타인",
     enNation: "Liechtenstein",
   },
   {
+    code: "MA",
     krNation: "마다가스카르",
     enNation: "Madagascar",
   },
   {
+    code: "MA",
     krNation: "마르티니크",
     enNation: "Martinique",
   },
   {
+    code: "MA",
     krNation: "마셜 제도",
     enNation: "Marshall Islands",
   },
   {
+    code: "MA",
     krNation: "마요트",
     enNation: "Mayotte",
   },
   {
+    code: "MA",
     krNation: "마카오",
     enNation: "Macao",
   },
   {
+    code: "MA",
     krNation: "말라위",
     enNation: "Malawi",
   },
   {
+    code: "MA",
     krNation: "말레이시아",
     enNation: "Malaysia",
   },
   {
+    code: "MA",
     krNation: "말리",
     enNation: "Mali",
   },
   {
+    code: "ME",
     krNation: "멕시코",
     enNation: "Mexico",
   },
   {
+    code: "MO",
     krNation: "모나코",
     enNation: "Monaco",
   },
   {
+    code: "MO",
     krNation: "모로코",
     enNation: "Morocco",
   },
   {
+    code: "MA",
     krNation: "모리셔스",
     enNation: "Mauritius",
   },
   {
+    code: "MA",
     krNation: "모리타니",
     enNation: "Mauritania",
   },
   {
+    code: "MO",
     krNation: "모잠비크",
     enNation: "Mozambique",
   },
   {
+    code: "MO",
     krNation: "몬테네그로",
     enNation: "Montenegro",
   },
   {
+    code: "MO",
     krNation: "몬트세랫",
     enNation: "Montserrat",
   },
   {
+    code: "MA",
     krNation: "몰도바",
     enNation: "Maldova",
   },
   {
+    code: "MA",
     krNation: "몰디브",
     enNation: "Maldives",
   },
   {
+    code: "MA",
     krNation: "몰타",
     enNation: "Malta",
   },
   {
+    code: "MO",
     krNation: "몽골",
     enNation: "Mongolia",
   },
   {
+    code: "UN",
     krNation: "미국",
     enNation: "United States of America",
   },
   {
+    code: "MY",
     krNation: "미얀마",
     enNation: "Myanmar",
   },
   {
+    code: "MI",
     krNation: "미크로네시아 연방",
     enNation: "Micronesia",
   },
   {
+    code: "VA",
     krNation: "바누아투",
     enNation: "Vanuatu",
   },
   {
+    code: "BA",
     krNation: "바레인",
     enNation: "Bahrain",
   },
   {
+    code: "BA",
     krNation: "바베이도스",
     enNation: "Barbados",
   },
   {
+    code: "VA",
     krNation: "바티칸 시국",
     enNation: "Vatican",
   },
   {
+    code: "BA",
     krNation: "바하마",
     enNation: "Bahamas",
   },
   {
+    code: "BA",
     krNation: "방글라데시",
     enNation: "Bangladesh",
   },
   {
+    code: "BE",
     krNation: "버뮤다",
     enNation: "Bermuda",
   },
   {
+    code: "BE",
     krNation: "베냉",
     enNation: "Benin",
   },
   {
+    code: "VE",
     krNation: "베네수엘라",
     enNation: "Venezuela",
   },
   {
+    code: "VI",
     krNation: "베트남",
     enNation: "Vietnam",
   },
   {
+    code: "BE",
     krNation: "벨기에",
     enNation: "Belgium",
   },
   {
+    code: "BE",
     krNation: "벨라루스",
     enNation: "Belarus",
   },
   {
+    code: "BE",
     krNation: "벨리즈",
     enNation: "Belize",
   },
   {
+    code: "BO",
     krNation: "보스니아 헤르체고비나",
     enNation: "Bosnia and Herzegovina",
   },
   {
+    code: "BO",
     krNation: "보츠와나",
     enNation: "Botswana",
   },
   {
+    code: "BO",
     krNation: "볼리비아",
     enNation: "Bolivia",
   },
   {
+    code: "BU",
     krNation: "부룬디",
     enNation: "Burundi",
   },
   {
+    code: "BU",
     krNation: "부르키나파소",
     enNation: "Burkina Faso",
   },
   {
+    code: "BH",
     krNation: "부탄",
     enNation: "Bhutan",
   },
   {
+    code: "MA",
     krNation: "북마케도니아",
     enNation: "Macedonia",
   },
   {
+    code: "BU",
     krNation: "불가리아",
     enNation: "Bulgaria",
   },
   {
+    code: "BR",
     krNation: "브라질",
     enNation: "Brazil",
   },
   {
+    code: "BR",
     krNation: "브루나이",
     enNation: "Brunei",
   },
   {
+    code: "SA",
     krNation: "사모아",
     enNation: "Samoa",
   },
   {
+    code: "SA",
     krNation: "사우디아라비아",
     enNation: "Saudi Arabia",
   },
   {
+    code: "SA",
     krNation: "산마리노",
     enNation: "San Marino",
   },
   {
+    code: "SA",
     krNation: "생피에르 미클롱",
     enNation: "Saint Pierre and Miquelon",
   },
   {
+    code: "WE",
     krNation: "서사하라",
     enNation: "Western Sahara",
   },
   {
+    code: "SE",
     krNation: "세네갈",
     enNation: "Senegal",
   },
   {
+    code: "SE",
     krNation: "세르비아",
     enNation: "Serbia",
   },
   {
+    code: "SE",
     krNation: "세이셸",
     enNation: "Seychelles",
   },
   {
+    code: "SA",
     krNation: "세인트루시아",
     enNation: "Saint Lucia",
   },
   {
+    code: "SA",
     krNation: "세인트헬레나",
     enNation: "Saint Helena",
   },
   {
+    code: "SO",
     krNation: "소말리아",
     enNation: "Somalia",
   },
   {
+    code: "SO",
     krNation: "솔로몬 제도",
     enNation: "Solomon Islands",
   },
   {
+    code: "SU",
     krNation: "수단",
     enNation: "Sudan",
   },
   {
+    code: "SU",
     krNation: "수리남",
     enNation: "Suriname",
   },
   {
+    code: "SR",
     krNation: "스리랑카",
     enNation: "Sri Lanka",
   },
   {
+    code: "SW",
     krNation: "스웨덴",
     enNation: "Sweden",
   },
   {
+    code: "SW",
     krNation: "스위스",
     enNation: "Switzerland",
   },
   {
+    code: "SP",
     krNation: "스페인",
     enNation: "Spain",
   },
   {
+    code: "SL",
     krNation: "슬로바키아",
     enNation: "Slovakia",
   },
   {
+    code: "SL",
     krNation: "슬로베니아",
     enNation: "Slovenia",
   },
   {
+    code: "SY",
     krNation: "시리아",
     enNation: "Syria",
   },
   {
+    code: "SI",
     krNation: "시에라리온",
     enNation: "Sierra Leone",
   },
   {
+    code: "SI",
     krNation: "신트마르턴",
     enNation: "Sint Maarten",
   },
   {
+    code: "SI",
     krNation: "싱가포르",
     enNation: "Singapore",
   },
   {
+    code: "UN",
     krNation: "아랍에미리트",
     enNation: "United Arab Emirates",
   },
   {
+    code: "AR",
     krNation: "아루바",
     enNation: "Aruba",
   },
   {
+    code: "AR",
     krNation: "아르메니아",
     enNation: "Armenia",
   },
   {
+    code: "AR",
     krNation: "아르헨티나",
     enNation: "Argentina",
   },
   {
+    code: "AM",
     krNation: "아메리칸사모아",
     enNation: "American Samoa",
   },
   {
+    code: "IC",
     krNation: "아이슬란드",
     enNation: "Iceland",
   },
   {
+    code: "HA",
     krNation: "아이티",
     enNation: "Haiti",
   },
   {
+    code: "IR",
     krNation: "아일랜드",
     enNation: "Ireland",
   },
   {
+    code: "AZ",
     krNation: "아제르바이잔",
     enNation: "Azerbaijan",
   },
   {
+    code: "AF",
     krNation: "아프가니스탄",
     enNation: "Afghanistan",
   },
   {
+    code: "AN",
     krNation: "안도라",
     enNation: "Andorra",
   },
   {
+    code: "AL",
     krNation: "알바니아",
     enNation: "Albania",
   },
   {
+    code: "AL",
     krNation: "알제리",
     enNation: "Algeria",
   },
   {
+    code: "AN",
     krNation: "앙골라",
     enNation: "Angola",
   },
   {
+    code: "AN",
     krNation: "앤티가 바부다",
     enNation: "Antigua and Barbuda",
   },
   {
+    code: "AN",
     krNation: "앵귈라",
     enNation: "Anguilla",
   },
   {
+    code: "ER",
     krNation: "에리트레아",
     enNation: "Eritrea",
   },
   {
+    code: "ES",
     krNation: "에스와티니",
     enNation: "Eswatini",
   },
   {
+    code: "ES",
     krNation: "에스토니아",
     enNation: "Estonia",
   },
   {
+    code: "EC",
     krNation: "에콰도르",
     enNation: "Ecuador",
   },
   {
+    code: "ET",
     krNation: "에티오피아",
     enNation: "Ethiopia",
   },
   {
+    code: "EL",
     krNation: "엘살바도르",
     enNation: "El Salvador",
   },
   {
+    code: "UN",
     krNation: "영국",
     enNation: "United Kingdom",
   },
   {
+    code: "YE",
     krNation: "예멘",
     enNation: "Yemen",
   },
   {
+    code: "OM",
     krNation: "오만",
     enNation: "Oman",
   },
   {
+    code: "AU",
     krNation: "오스트레일리아",
     enNation: "Australia",
   },
   {
+    code: "AU",
     krNation: "오스트리아",
     enNation: "Austria",
   },
   {
+    code: "HO",
     krNation: "온두라스",
     enNation: "Honduras",
   },
   {
+    code: "JO",
     krNation: "요르단",
     enNation: "Jordan",
   },
   {
+    code: "UG",
     krNation: "우간다",
     enNation: "Uganda",
   },
   {
+    code: "UR",
     krNation: "우루과이",
     enNation: "Uruguay",
   },
   {
+    code: "UZ",
     krNation: "우즈베키스탄",
     enNation: "Uzbekistan",
   },
   {
+    code: "UK",
     krNation: "우크라이나",
     enNation: "Ukraine",
   },
   {
+    code: "IR",
     krNation: "이라크",
     enNation: "Iraq",
   },
   {
+    code: "IR",
     krNation: "이란",
     enNation: "Iran",
   },
   {
+    code: "IS",
     krNation: "이스라엘",
     enNation: "Israel",
   },
   {
+    code: "EG",
     krNation: "이집트",
     enNation: "Egypt",
   },
   {
+    code: "IT",
     krNation: "이탈리아",
     enNation: "Italy",
   },
   {
+    code: "IN",
     krNation: "인도",
     enNation: "India",
   },
   {
+    code: "IN",
     krNation: "인도네시아",
     enNation: "Indonesia",
   },
   {
+    code: "JA",
     krNation: "일본",
     enNation: "Japan",
   },
   {
+    code: "JA",
     krNation: "자메이카",
     enNation: "Jamaica",
   },
   {
+    code: "ZA",
     krNation: "잠비아",
     enNation: "Zambia",
   },
   {
+    code: "EQ",
     krNation: "적도 기니",
     enNation: "Equatorial Guinea",
   },
   {
-    krNation: "북한",
-    enNation: "DPR of Korea",
-  },
-  {
+    code: "GE",
     krNation: "조지아",
     enNation: "Georgia",
   },
   {
+    code: "CE",
     krNation: "중앙아프리카 공화국",
     enNation: "Central Africa",
   },
   {
+    code: "TA",
     krNation: "대만",
     enNation: "Taiwan",
   },
   {
+    code: "CH",
     krNation: "중화인민공화국",
     enNation: "China",
   },
   {
+    code: "DJ",
     krNation: "지부티",
     enNation: "Djibouti",
   },
   {
+    code: "GI",
     krNation: "지브롤터",
     enNation: "Gibraltar",
   },
   {
+    code: "ZI",
     krNation: "짐바브웨",
     enNation: "Zimbabwe",
   },
   {
+    code: "CH",
     krNation: "차드",
     enNation: "Chad",
   },
   {
+    code: "CZ",
     krNation: "체코",
     enNation: "Czech",
   },
   {
+    code: "CH",
     krNation: "칠레",
     enNation: "Chile",
   },
   {
+    code: "CA",
     krNation: "카메룬",
     enNation: "Cameroon",
   },
   {
+    code: "CA",
     krNation: "카보베르데",
     enNation: "Cape Verde",
   },
   {
+    code: "KA",
     krNation: "카자흐스탄",
     enNation: "Kazakstan",
   },
   {
+    code: "QA",
     krNation: "카타르",
     enNation: "Qatar",
   },
   {
+    code: "CA",
     krNation: "캄보디아",
     enNation: "Cambodia",
   },
   {
+    code: "CA",
     krNation: "캐나다",
     enNation: "Canada",
   },
   {
+    code: "KE",
     krNation: "케냐",
     enNation: "Kenya",
   },
   {
+    code: "CA",
     krNation: "케이맨 제도",
     enNation: "Cayman Islands",
   },
   {
+    code: "TH",
     krNation: "코모로",
     enNation: "the Comoros",
   },
   {
+    code: "CO",
     krNation: "코스타리카",
     enNation: "Costa Rica",
   },
   {
+    code: "IV",
     krNation: "코트디부아르",
     enNation: "Ivory Coast",
   },
   {
+    code: "CO",
     krNation: "콜롬비아",
     enNation: "Colombia",
   },
   {
+    code: "CO",
     krNation: "콩고 공화국",
     enNation: "Congo",
   },
   {
+    code: "DR",
     krNation: "콩고 민주 공화국",
     enNation: "DR Congo",
   },
   {
+    code: "CU",
     krNation: "쿠바",
     enNation: "Cuba",
   },
   {
+    code: "KU",
     krNation: "쿠웨이트",
     enNation: "Kuwait",
   },
   {
+    code: "CO",
     krNation: "쿡 제도",
     enNation: "Cook Islands",
   },
   {
+    code: "CR",
     krNation: "크로아티아",
     enNation: "Croatia",
   },
   {
+    code: "KY",
     krNation: "키르기스스탄",
     enNation: "Kyrgizstan",
   },
   {
+    code: "KI",
     krNation: "키리바시",
     enNation: "Kiribati",
   },
   {
+    code: "CY",
     krNation: "키프로스",
     enNation: "Cyprus",
   },
   {
+    code: "TA",
     krNation: "타지키스탄",
     enNation: "Tajikistan",
   },
   {
+    code: "TA",
     krNation: "탄자니아",
     enNation: "Tanzania",
   },
   {
+    code: "TH",
     krNation: "태국",
     enNation: "Thailand",
   },
   {
+    code: "TU",
     krNation: "터키",
     enNation: "Turkey",
   },
   {
+    code: "TO",
     krNation: "토고",
     enNation: "Togo",
   },
   {
+    code: "TO",
     krNation: "토켈라우",
     enNation: "Tokelau",
   },
   {
+    code: "TO",
     krNation: "통가",
     enNation: "Tonga",
   },
   {
+    code: "TU",
     krNation: "투르크메니스탄",
     enNation: "Turkmenistan",
   },
   {
+    code: "TU",
     krNation: "투발루",
     enNation: "Tuvalu",
   },
   {
+    code: "TU",
     krNation: "튀니지",
     enNation: "Tunisia",
   },
   {
+    code: "TR",
     krNation: "트리니다드 토바고",
     enNation: "Trinidad and Tobago",
   },
   {
+    code: "PA",
     krNation: "파나마",
     enNation: "Panama",
   },
   {
+    code: "PA",
     krNation: "파라과이",
     enNation: "Paraguay",
   },
   {
+    code: "PA",
     krNation: "파키스탄",
     enNation: "Pakistan",
   },
   {
+    code: "PA",
     krNation: "파푸아뉴기니",
     enNation: "Papua New Guinea",
   },
   {
+    code: "PA",
     krNation: "팔라우",
     enNation: "Palau",
   },
   {
+    code: "ST",
     krNation: "팔레스타인",
     enNation: "State of Palestine",
   },
   {
+    code: "PE",
     krNation: "페루",
     enNation: "Peru",
   },
   {
+    code: "PO",
     krNation: "포르투갈",
     enNation: "Portugal",
   },
   {
+    code: "FA",
     krNation: "포클랜드 제도",
     enNation: "Falkland Islands",
   },
   {
+    code: "PO",
     krNation: "폴란드",
     enNation: "Poland",
   },
   {
+    code: "PU",
     krNation: "푸에르토리코",
     enNation: "Puerto Rico",
   },
   {
+    code: "FR",
     krNation: "프랑스",
     enNation: "France",
   },
   {
+    code: "FI",
     krNation: "피지",
     enNation: "Fiji",
   },
   {
+    code: "FI",
     krNation: "핀란드",
     enNation: "Finland",
   },
   {
+    code: "TH",
     krNation: "필리핀",
     enNation: "the Philippines",
   },
   {
+    code: "HU",
     krNation: "헝가리",
     enNation: "Hungary",
   },
   {
+    code: "HO",
     krNation: "홍콩",
     enNation: "Hong Kong",
   },
