--- src/types/list.ts
@@ -1,8 +1,0 @@
-export type CountryListType = {
-  krNation: string;
-  enNation: string;
-};
-
-export type LanguageListType = {
-  languageList: string;
-};
